requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# PythonProject/main.py: 5
Pillow == 8.1.2

# PythonProject/GUI.py: 2
# PythonProject/canvasWidget.py: 3,4
# PythonProject/chartWidget.py: 3,4
PyQt5_sip == 12.8.1

# PythonProject/GUI.py: 1
# PythonProject/canvasWidget.py: 5
# PythonProject/main.py: 4
numpy == 1.20.1

# PythonProject/main.py: 3
# PythonProject/neuralNetwork.py: 1,3,4,5
torch == 1.8.1

# PythonProject/neuralNetwork.py: 2
torchvision == 0.9.1
